CASSANDRA-17964 3.11
patch by Stefan Miklosovic; reviewed by Berenguer Blasi and David Capwell for CASSANDRA-17943
