+++ test/distributed/org/apache/cassandra/distributed/test/ReprepareOldBehaviourTest.java
@@ -30,7 +30,7 @@ import static org.apache.cassandra.distributed.api.Feature.GOSSIP;
 import static org.apache.cassandra.distributed.api.Feature.NATIVE_PROTOCOL;
 import static org.apache.cassandra.distributed.api.Feature.NETWORK;
 
-public class ReprepareTestOldBehaviour extends ReprepareTestBase
+public class ReprepareOldBehaviourTest extends ReprepareTestBase
 {
     @Test
     public void testReprepareMixedVersion() throws Throwable
@@ -77,7 +77,7 @@ public class ReprepareTestOldBehaviour extends ReprepareTestBase
         }
     }
 
-    @Test
+    @Test // see CASSANDRA-18021
     public void testReprepareMixedVersionWithoutReset() throws Throwable
     {
         try (ICluster<IInvokableInstance> c = init(builder().withNodes(2)
